extract hcx valdation buttons to it's own component
